Face Recognition folder
This folder consists of face dectection and recognition codes in python. I have created a datasets folder to take images from my PC camera and save it in GRAY scale, then trained this data to learn about the person. after that a face recognition model using openCV library recognises the detected face based on the data it trained on.